Reorganized mesh reading
